+++ test/webdev_server_test.dart
@@ -13,8 +13,6 @@
 // limitations under the License.
 
 @TestOn('vm')
-import 'dart:io';
-
 import 'package:http/http.dart' as http;
 import 'package:test/test.dart';
 import 'package:webdev_proxy/src/port_utils.dart';
@@ -36,8 +34,7 @@ void main() async {
 
   test('Serves a directory', () async {
     final port = await findUnusedPort();
-    webdevServer =
-        await WebdevServer.start(['test:$port'], mode: ProcessStartMode.normal);
+    webdevServer = await WebdevServer.start(['test:$port']);
 
     // We don't have a good way of knowing when the `webdev serve` process has
     // started listening on the port, so we send a request periodically until it